ended up splitting test files anyway due to size of builder tests
