+++ tests/obv3_json_examples/alignment_case.json
