+++ tests/obv3_json_examples/endorsement_credential.json
