+++ tests/obv3_json_examples/skill_assertion_case.json
